MYM-118 pingpong logic fixed.
