+++ frontend/front/src/pages/GamePage.tsx
@@ -32,27 +32,20 @@ export default function GamePage() {
 
   const [adminConsole, setAdminConsole] = useAtom(chatAtom.adminConsoleAtom);
 
-  const [serverClientTimeDiff, setServerClientTimeDiff] = useAtom(serverClientTimeDiffAtom);
-
-  let pingInterval: NodeJS.Timer;
-
-  // 1 sec delay for init value
-  let pingRTTmin: number = 2000;
-
   PressKey(["F4"], () => {
     setAdminConsole((prev) => !prev);
   });
 
   useEffect(() => {
     if (isLoading === false) {
-      AdminLogPrinter(adminConsole, "gameSocket connection");
+      AdminLogPrinter(adminConsole, `gameSocket connection`);
       game.gameSocket.connect();
       setIsLoading(true);
+      return () => {
+        game.gameSocket.disconnect();
+      };
     }
-    return () => {
-      game.gameSocket.disconnect();
-    };
-  }, [game.gameSocket]);
+  }, []);
 
   const connectionEventHandler = () => {
     if (game.gameSocket.connected) {
@@ -62,7 +55,7 @@ export default function GamePage() {
       } else {
         // new or unrecoverable session
         AdminLogPrinter(adminConsole, "gameSocket connected");
-        pingInterval = setInterval(pingEvent, 1000);
+        // pingInterval = setInterval(pingEvent, 1000);
       }
     }
   };
@@ -71,38 +64,23 @@ export default function GamePage() {
   const disconnectionEventHandler = (reason: string) => {
     if (reason === "io server disconnect") {
     }
-    clearInterval(pingInterval);
-    // setIsQueue(false);
-    setIsLoading(false);
-    setIsPrivate(false);
+    if (isLoading) {
+      setIsLoading(false);
+    }
+    if (isPrivate) {
+      setIsPrivate(false);
+    }
     AdminLogPrinter(adminConsole, "gameSocket disconnected");
-  };
-
-  const pingEvent = () => {
-    const curTime = Date.now();
-    const pingEventHandler = (serverTime: number) => {
-      const now = Date.now();
-      const pingRTT = now - curTime;
-      AdminLogPrinter(adminConsole, `pingRTT: ${pingRTT}ms`);
-      if (pingRTTmin > pingRTT) {
-        pingRTTmin = pingRTT;
-        const adjServerTime = serverTime + pingRTTmin / 2;
-        AdminLogPrinter(adminConsole, `updated serverClientTimeDiff: ${serverClientTimeDiff}ms`);
-        setServerClientTimeDiff(now - adjServerTime);
-        AdminLogPrinter(adminConsole, `updated serverClientTimeDiff: ${serverClientTimeDiff}ms`);
-      }
-      AdminLogPrinter(adminConsole, `pingRTTmin: ${pingRTTmin}ms`);
-    };
-    game.gameSocket.emit("ping", pingEventHandler);
-    return () => {
-      game.gameSocket.off("ping", pingEventHandler);
-    };
   };
 
   const startEventHandler = () => {
     AdminLogPrinter(adminConsole, "game start");
-    setIsLoading(false);
-    setIsGameStart(true);
+    if (isLoading) {
+      setIsLoading(false);
+    }
+    if (!isGameStart) {
+      setIsGameStart(true);
+    }
   };
 
   useEffect(() => {
@@ -113,19 +91,17 @@ export default function GamePage() {
   }, [isLoading, isPrivate, isGameStart]);
 
   useEffect(() => {
-    game.gameSocket.on("connect", connectionEventHandler);
-    game.gameSocket.on("disconnect", disconnectionEventHandler);
-    game.gameSocket.on("gameStart", startEventHandler);
-    // game.gameSocket.on("isQueue", queueEventHandler);
-    // game.gameSocket.on("isLoading", loadingEventHandler);
-    return () => {
-      game.gameSocket.off("connect", connectionEventHandler);
-      game.gameSocket.off("disconnect", disconnectionEventHandler);
-      game.gameSocket.off("gameStart", startEventHandler);
-      // game.gameSocket.off("isQueue", queueEventHandler);
-      // game.gameSocket.off("isLoading", loadingEventHandler);
-    };
-  }, [isLoading, isPrivate, isGameStart]);
+    if (isLoading) {
+      game.gameSocket.on("connect", connectionEventHandler);
+      game.gameSocket.on("disconnect", disconnectionEventHandler);
+      game.gameSocket.on("gameStart", startEventHandler);
+      return () => {
+        game.gameSocket.off("connect", connectionEventHandler);
+        game.gameSocket.off("disconnect", disconnectionEventHandler);
+        game.gameSocket.off("gameStart", startEventHandler);
+      };
+    }
+  }, [isLoading]);
 
   return (
     <BackGround>
